+++ src/agent/samplers/mod.rs
@@ -8,6 +8,7 @@ use std::sync::Arc;
 mod blockio;
 mod cpu;
 mod drivehealth;
+mod filesystem;
 mod gpu;
 mod memory;
 mod network;